refactor: solve conflicts, rename mocks folder, move state code blocks in Item and Main components to separate files
