+++ src/pages/Item/Data/types.ts
@@ -1,0 +1,5 @@
+import type { Character } from '#types/interfaces';
+
+export type DataProps = {
+  data: Character;
+};
