+++ src/pages/Item/Error/Error.module.scss
@@ -1,0 +1,7 @@
+.errorMessage {
+  padding: 1rem;
+  background-color: #fee2e2;
+  color: #b91c1c;
+  border-radius: 0.375rem;
+  border: 0.063rem solid #fca5a5;
+}
