+++ src/pages/Item/Error/types.ts
@@ -1,0 +1,6 @@
+import type { SerializedError } from '@reduxjs/toolkit';
+import type { FetchBaseQueryError } from '@reduxjs/toolkit/query';
+
+export type ErrorProps = {
+  error: FetchBaseQueryError | SerializedError;
+};
