+++ src/pages/Item/Item.module.scss
@@ -37,51 +37,10 @@
   padding-bottom: 0.5rem;
 }
 
-.spinnerContainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 10rem;
-}
-
-.errorMessage {
-  padding: 1rem;
-  background-color: #fee2e2;
-  color: #b91c1c;
-  border-radius: 0.375rem;
-  border: 0.063rem solid #fca5a5;
-}
-
 .emptyState {
   text-align: center;
   padding: 2rem 0;
   color: #6b7280;
-}
-
-.content {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.name {
-  font-size: 1.125rem;
-  font-weight: 700;
-}
-
-.detailsGrid {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 1rem;
-}
-
-.detailLabel {
-  font-size: 0.875rem;
-  color: #6b7280;
-}
-
-.detailValue {
-  font-weight: 500;
 }
 
 .header {
