+++ src/pages/Item/Loading/Loading.module.scss
@@ -1,0 +1,6 @@
+.spinnerContainer {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 10rem;
+}
