+++ src/pages/Item/__tests__/index.test.tsx
@@ -6,19 +6,21 @@ import {
 } from '@testing-library/react';
 import { http, HttpResponse } from 'msw';
 import { createMemoryRouter, RouterProvider } from 'react-router';
-import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 
+import { defineGlobals } from '../../../../tests/setup';
+import { server } from '../../../__mocks__/server';
 import { MAIN_ROUTE } from '../../../constants';
-import { server } from '../../../mocks/server';
-import { routes } from '../../../router';
 import { routes } from '../../../router';
 import { StubProvider } from '../../../router/utils';
 import { renderWithProviders } from '../../../store/util';
 import Item from '..';
-import Item from '..';
 
 describe('Item page', () => {
+  beforeAll(() => {
+    defineGlobals();
+  });
+
   beforeEach(() => {
     vi.resetAllMocks();
     localStorage.clear();
