+++ src/pages/Main/Main.module.scss
@@ -43,14 +43,6 @@
   font-weight: 600;
 }
 
-.errorMessage {
-  padding: 1rem;
-  background-color: #fee2e2;
-  color: #b91c1c;
-  border-radius: 0.375rem;
-  border: 0.063rem solid #fca5a5;
-}
-
 .selectionPanel {
   position: fixed;
   z-index: 10;
