+++ tests/setup.ts
@@ -4,7 +4,7 @@ import { afterEach } from 'node:test';
 
 import { afterAll, beforeAll, vi } from 'vitest';
 
-import { server } from '../src/mocks/server';
+import { server } from '../src/__mocks__/server';
 
 export const mockCharacter = {
   name: 'Luke Skywalker',
